% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_code.R
\name{checkIDs}
\alias{checkIDs}
\title{Check matching IDs between metaxa2 taxonomy and fasta files}
\usage{
checkIDs(metax = NULL, seq = NULL)
}
\arguments{
\item{metax}{File name: the metaxa2 taxonomy file}

\item{seq}{The metaxa2 sequence file (fasta format)}
}
\description{
Check matching IDs between metaxa2 taxonomy and fasta files
}
