% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_code.R
\name{combine_files}
\alias{combine_files}
\title{Combine multiple fasta or taxonomy files into one}
\usage{
combine_files(list = NULL, out = NULL)
}
\arguments{
\item{list}{A character vector of files that you want to combine}

\item{out}{the output combined file name}
}
\description{
Combine multiple fasta or taxonomy files into one
}
