% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_code.R
\name{make_db}
\alias{make_db}
\title{Make a metaxa2 database}
\usage{
make_db(fasta = NULL, tax = NULL, outdir = NULL, name = NULL)
}
\arguments{
\item{fasta}{The metaxa2 sequence data (fasta file)}

\item{tax}{The metaxa2 taxonomy data}

\item{outdir}{The directory to save the database into}

\item{name}{The name to give to the database}
}
\description{
Make a metaxa2 database
}
