% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_code.R
\name{metax2tax}
\alias{metax2tax}
\title{Convert from Metaxa taxonomy format to a tabular, easy-to-read format}
\usage{
metax2tax(metax = NULL, out = NULL)
}
\arguments{
\item{metax}{File name: the file you wish to convert}

\item{out}{File name: the output file}
}
\description{
Convert from Metaxa taxonomy format to a tabular, easy-to-read format
}
