% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_code.R
\name{run_metaxa2}
\alias{run_metaxa2}
\title{Run metaxa2 on a dataset}
\usage{
run_metaxa2(db = NULL, input = NULL, out = NULL)
}
\arguments{
\item{db}{location of the directory containing the metaxa2 database}

\item{input}{input query fasta file}

\item{out}{base name for output files (including folder location)}
}
\description{
Run metaxa2 on a dataset
}
