% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_code.R
\name{setup_package}
\alias{setup_package}
\title{Execute this function the first time you install the package}
\usage{
setup_package()
}
\description{
Execute this function the first time you install the package
}
