% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_code.R
\name{tax2metax}
\alias{tax2metax}
\title{Convert from tabular taxonomy file to a metaxa2-format taxonomy file}
\usage{
tax2metax(tax = NULL, out = NULL)
}
\arguments{
\item{tax}{File name: the file you wish to convert}

\item{out}{File name: the output file}
}
\description{
Convert from tabular taxonomy file to a metaxa2-format taxonomy file
}
