% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_code.R
\name{unique_seqs}
\alias{unique_seqs}
\title{Create a fasta file containing only unique sequences from your data}
\usage{
unique_seqs(list = NULL, outdir = NULL)
}
\arguments{
\item{list}{A character vector of fasta files that you want to include}

\item{outdir}{The folder location in which to write the results}
}
\description{
The function also outputs the file "correspondence.tab", which details the correspondence between original sequences and the defined unique sequences.
}
\details{
It also outputs the file "cross_table.tab", which summarises the number of each unique sequence in each file.
}
